requirements.txt
torch
torchvision
PyYAML
numpy
Pillow
openpyxl
